[FIX]: Formating the datas according to the expectations (temp to 0.1 - press & humi to int)

+++ datasMeteo.h
@@ -9,8 +9,8 @@ private :
     QString picto;                      // icone pour les données
     QString ville;                      // la ville associée aux donnée tempérées
     double temperatureCelsius = 0.0;    // temperature en celsius
-    double humidity = 0.0;              // humitidty en %
-    double pressure = 0.0;              // pressure en hPa
+    int humidity = 0;              // humitidty en %
+    int pressure = 0;              // pressure en hPa
     QString date;
 
 
@@ -26,8 +26,8 @@ public:
     double getTemperatureCelsius() const;
     double getTemperatureKelvin() const;
     double getTemperatureFahrenheit() const;
-    double getHumidity() const;
-    double getPressure() const;
+    int getHumidity() const;
+    int getPressure() const;
 
     void setDate(const QString& vdate);
     void setPicto(const QString& vPicto);
@@ -35,8 +35,8 @@ public:
     void setTemperatureCelsius(const double& vTemperature);
     void setTemperatureKelvin(const double& vTemperature);
     void setTemperatureFahrenheit(const double& vTemperature);
-    void setHumidity(const double& vHumidity);
-    void setPressure(const double& vPressure);
+    void setHumidity(const int& vHumidity);
+    void setPressure(const int& vPressure);
 };
 
 #endif
